+++ packages/terradart_google/lib/secret_manager.dart
@@ -4,12 +4,14 @@ library;
 
 export 'src/secret_manager/google_secret_manager_secret.dart'
     show
+        SecretManagerSecretAutoReplication,
         SecretManagerSecretCustomerManagedEncryption,
         GoogleSecretManagerSecret,
         SecretManagerSecretReplica,
         SecretManagerSecretReplication,
         SecretManagerSecretRotation,
-        SecretManagerSecretSecretTopic;
+        SecretManagerSecretSecretTopic,
+        SecretManagerSecretUserManagedReplication;
 export 'src/secret_manager/google_secret_manager_secret_iam_member.dart'
     show GoogleSecretManagerSecretIamMember;
 export 'src/secret_manager/google_secret_manager_secret_version.dart'